requirements.txt
GitPython==3.1.40
pandas==2.1.4
matplotlib==3.8.2
seaborn==0.13.0
numpy==1.26.2
plotly==5.17.0
